backend/requirements.txt
amqp==5.3.1
asgiref==3.10.0
billiard==4.2.2
celery==5.5.3
sendgrid==6.12.5
click==8.3.0
click-didyoumean==0.3.1
click-plugins==1.1.1.2
click-repl==0.3.0
Django==5.2.7
django-cors-headers==4.9.0
djangorestframework==3.16.1
djangorestframework_simplejwt==5.5.1
drf-yasg==1.21.11
ecdsa==0.19.1
inflection==0.5.1
kombu==5.5.4
packaging==25.0
prompt_toolkit==3.0.52
psycopg2-binary==2.9.11
pyasn1==0.6.1
PyJWT==2.10.1
python-dateutil==2.9.0.post0
python-dotenv==1.1.1
python-jose==3.5.0
pytz==2025.2
PyYAML==6.0.3
redis==6.4.0
rsa==4.9.1
six==1.17.0
sqlparse==0.5.3
tzdata==2025.2
drf-spectacular==0.28.0
uritemplate==4.2.0
vine==5.1.0
wcwidth==0.2.14
requests